Rename qdrant_client module to not clash with qdrant_client library
