+++ tests/load/qdrant/test_pipeline.py
@@ -10,7 +10,7 @@ from dlt.common.typing import DictStrStr
 
 from dlt.destinations.adapters import qdrant_adapter
 from dlt.destinations.impl.qdrant.qdrant_adapter import qdrant_adapter, VECTORIZE_HINT
-from dlt.destinations.impl.qdrant.qdrant_client import QdrantClient
+from dlt.destinations.impl.qdrant.qdrant_job_client import QdrantClient
 from tests.pipeline.utils import assert_load_info
 from tests.load.qdrant.utils import drop_active_pipeline_data, assert_collection
 from tests.load.utils import sequence_generator
